CNN_tranining/requirements.txt
astroid==2.4.2
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
flickrapi==2.4.0
future==0.18.2
gapcv==1.0.0
google-images-download==2.8.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0      
isort==5.6.4
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
networkx==2.5
numpy==1.19.2
oauthlib==3.1.0
opencv-contrib-python==4.4.0.44
opencv-python==4.4.0.44
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.0
prompt-toolkit==3.0.8
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
pywin32==228
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
scikit-image==0.17.2
scipy==1.5.3
selenium==3.141.0
six==1.15.0
tifffile==2020.10.1
toml==0.10.1
torch
torchsummary==1.5.1
torchvision
tornado==6.0.4
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
wrapt==1.12.1